glih-backend/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glih-backend"
version = "0.1.0"
description = "GLIH FastAPI backend"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Lanre" }]
license = { text = "Proprietary" }
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.23.0",
  "python-dotenv>=1.0.1",
  "tomli>=2.0.1",
  "openai>=1.0.0",
  "chromadb>=0.4.22",
  "requests>=2.31.0",
  "tomli-w>=1.0.0",
  "pypdf>=3.9.0",
  "python-multipart>=0.0.6",
  "beautifulsoup4>=4.12.0",
  "pdfminer.six>=20221105",
]

[project.optional-dependencies]
dev = ["pytest>=7", "httpx>=0.24"]
llms = [
  "anthropic>=0.37.0",
  "mistralai>=1.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"]}}
include = ["glih_backend*"]

[project.scripts]
glih-api = "glih_backend.__main__:main"
